New fitting limits, bugfixes, kind test code
fit_kind was revised to use new limits on the fitting process, after I noticed that the t_Pmax value could sometimes overrun the start of isovolumic contraction (seems bad to allow this?).

There were a few bugs in runAll (diary), fit_kind (needed a ...), and data_kind (needed a second arg to p0M).

fit_kind_test is a test code for looking at how the fit varies with ICs and bounds.