Reworking data_isoidx
Vanderpool indices in this form (using extrema funtions, min/max) don't seem to be working. Need to traverse the data like with Takeuchi and find the first change in direction.

For now, this commit saves this somewhat dysfunctional method (that worked... kind of) as well as the splitting of the three indices finding sections into subfunctions to make things easier to handle.

Next up will be fixing data_isoidx_v to use the same "data crawl" as data_isoidx_t does...