Generalized Cleaning out of Indices
The specific "cleaning" loops at the end of data_isoidx were generalized so they could be done by function call. This was really an effort to enable further cleaning to happen in data_isoseg, but that code has not yet been written. Perhaps put a zero check in that code, have it "error out" and if it's ever needed, I'll write it...