New Pes Finding, Bugfixes
This code encorporates the new (PA)max before (dP/dt)min method to find Pes, which is actually quite good.

It also fixes a MAJOR bug in the processing of waveforms that were not sampled at 250Hz - the stepsize was not correctly passed to data_centdiff so 6 or 7 of the Tedford files were badly handled.