Time-Normalized Kind
The original version of the Kind fit was not time-normalized (i.e. each cycle was fit in place, not normalized such that the cycle always starts at t=0). This worked, but yielded less consistent results (just like with Takeuchi). This is the normalized version. It is likely that there will be additional changes to remove the "_O" and "_N" (the non-normalized Takeuchi - Adam's - and the non-normalized Kind - Mine!) soon, but for now they can stay in the code...