Revision of Output Code
The output code worked fine enough, but every time something else was added, into the structure, the code required careful revision to the header (to align the output in the CSV) as well as to the actual fprintf statements. That system has been replaced by a code that simply plucks off the fields and prints them in alphabetical order (into the header and into the file).

One caveat may be that the header should be written out each time! Not just to a new file. But for now, I'll leave it as is; the header shouldn't change that much (only over revisions in the code, when RV_output.csv should be deleted anyway)...