Update test_calc_vvcr.m
This makes test_calc_vvcr spit out the first 3-7 values of the ivSeg.iv2Pres vectors. It showed that the landmarks were not being consistently obtained across the five runNum runs. Thus, the code needs to be revised to encoporate the entire "width" of the varying landmarks, then simply select the range using indices.

Next commit will be major, to revise as stated above.