Bugfix for double-hump Kind fit + plot bugfix
The Kind fit was sometimes returning "double hump" fits because its range for finding (dP/dt)min spaned an entire period... so sometimes it found the next one. The range was reduced (in both the fit and the plot routines) to only 70% of the period, so it never finds the 2nd minimum.

Also, a regexprep caused the ylabel in the Kind GUI to be funky (it was a variable name rather than "Pressure"). This was fixed.